+++ Assets/KBG/GameClear.cs
@@ -4,26 +4,22 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
-public class GameOverCL : MonoBehaviour
+public class GameClear : MonoBehaviour
 {
-    public static GameOverCL Instance;
-
-    public UnityEvent OnPlayerDied = new UnityEvent();
+    public static GameClear Instance;
 
     public UnityEvent OnGameClear = new UnityEvent();
 
-    public bool IsGameOvered;
     public bool IsGameCleared;
 
     [SerializeField] private GameObject Player;
 
     [Header("UI")]
-    [SerializeField] private GameObject PlayerDiedPanel;
     [SerializeField] private GameObject GameClearPanel;
 
     private void Awake()
     {
-        if(Instance == null)
+        if (Instance == null)
         {
             Instance = this;
             DontDestroyOnLoad(gameObject);
@@ -36,30 +32,17 @@ public class GameOverCL : MonoBehaviour
     private void OnEnable()
     {
         OnGameClear.AddListener(GameOver);
-        OnPlayerDied.AddListener(GameOver);
     }
-    
+
     public void GameOver()
-    {
-        Clear();
-        Fail();
-        
-    }
-    private void OnDisable()
-    {
-        OnGameClear.RemoveListener(GameOver);
-        OnPlayerDied.RemoveListener(GameOver);
-    }
-    public void Fail()
-    {
-        Player.SetActive(false);
-        IsGameOvered = true;
-        PlayerDiedPanel.SetActive(true);
-    }
-    public void Clear()
     {
         Player.SetActive(false);
         IsGameCleared = true;
         GameClearPanel.SetActive(true);
     }
+    private void OnDisable()
+    {
+        OnGameClear.RemoveListener(GameOver);
+    }
+    
 }
