+++ Assets/KBG/GameClear.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 00be58dbb8902c0449a3ba77bf1af624
+guid: 4d6e9a054d24bda4cb2e996807753e98
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
